backend/lesson_plan_generator/app/config/env.py
import os
from dotenv import load_dotenv # type: ignore

load_dotenv()

class Config:
    DEBUG = os.getenv("DEBUG", "False").lower() == "true"
    HOST = os.getenv("HOST", "0.0.0.0")
    PORT = int(os.getenv("PORT", 8005))
    CORS_ORIGINS = os.getenv("CORS_ORIGINS", "http://localhost:5173,http://127.0.0.1:5173").split(",")
    GEMINI_API_KEY = os.getenv("GEMINI_API_KEY", "")

config = Config()