backend/lesson_plan_generator/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
python-dotenv==1.0.0
google-generativeai==0.2.0